ディレクトリ構成の変更とCSS変更 /PrefecturesField/CheckBoxList/CheckBox ↓ /PrefecturesField/CheckBox
