--- src/app/components/PrefecturesField/CheckBoxList/CheckBoxList.module.css
@@ -1,19 +1,0 @@
-/* CheckBoxList.module.css */
-
-/* これは全体の外側のコンテナに適用されるスタイル */
-.container {
-  border: 2px solid gray; /* #ccc枠線のスタイルと色を指定 */
-  padding: 10px; /* 必要に応じてパディングを追加 */
-}
-
-/* 各地域のセクションに適用されるスタイル */
-.areaSection {
-  border-bottom: 2px solid gray; /* 各地域のセクションの下部に横線を追加 */
-  margin-bottom: 10px; /* 各セクションの下に余白を追加 */
-}
-
-/* タイトル（地域名）に適用されるスタイル */
-.areaTitle {
-  font-size: 18px; /* タイトルのフォントサイズを指定 */
-  margin-bottom: 5px; /* タイトルの下に余白を追加 */
-}
